Added double precision mode

+++ Computation/constants.h
@@ -1,3 +1,3 @@
 #pragma once
 
-constexpr float KERNEL_ATTRACTOR_MAX_TOL = 0.1f;
+constexpr double KERNEL_ATTRACTOR_MAX_TOL = 0.1;
